refactor(TransferManagerWorker): Cleaner Error declaration and encapsulation
