+++ SwissTransferFeatures/UploadProgressView/TransferManagerWorker+Error.swift
@@ -1,0 +1,31 @@
+/*
+ Infomaniak SwissTransfer - iOS App
+ Copyright (C) 2025 Infomaniak Network SA
+
+ This program is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import Foundation
+
+extension TransferManagerWorker {
+    enum ErrorDomain: Error {
+        case remoteContainerNotFound
+        case invalidURL(rawURL: String)
+        case invalidUploadChunkURL
+        case invalidChunk
+        case invalidRange
+        case invalidResponse
+        case invalidChunkResponse
+    }
+}
